pyproject.toml
[project]
name = "gourmet-wiki"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "django-formtools>=2.5.1",
]
